Create calculator.asm
Program an arithmetic calculator that can add, subtract, multiply and divide. Display proper messages on screen to ask user for the input and output numbers and the operation that needs to be performed. Incorporate procedures and macros in your code, e.g., you can make procedures for the arithmetic operations and macros for input and output.
--Input is taken using WinAPI calls